requirements.txt
astroid==2.4.2
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
colorama==0.4.4
dateparser==0.7.6
distro==1.5.0
Flask==1.1.2
gunicorn==20.0.4
idna==2.10
invoice2data==0.3.5
isort==5.6.4
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.19.4
pandas==1.1.4
pdfminer.six==20200517
pdfplumber==0.5.24
Pillow==8.0.1
pycryptodome==3.9.8
pylint==2.6.0
PyPDF2==1.26.0
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
regex==2020.10.23
requests==2.24.0
six==1.15.0
sortedcontainers==2.2.2
tabula-py==2.2.0
toml==0.10.2
tzlocal==2.1
Unidecode==1.1.1
urllib3==1.25.11
veryfi==0.0.4
Wand==0.6.3
Werkzeug==1.0.1
wincertstore==0.2
wrapt==1.12.1
